Simplify and reformat code
