+++ Makefile.am
@@ -9,12 +9,12 @@ libdir = $(PAMDIR)
 
 lib_LTLIBRARIES= pam_aad.la
 
-pam_aad_la_SOURCES = src/pam_aad.c src/jwt.c src/rest.c src/utils.c
+pam_aad_la_SOURCES = src/pam_aad.c src/rest.c src/utils.c
 pam_aad_la_LIBADD = -lcjson -lcrypto -ljwt -lm -lpam -lssl
 pam_aad_la_CFLAGS = $(AM_CFLAGS)
 pam_aad_la_LDFLAGS = $(AM_LDFLAGS) $(MODULES_LDFLAGS) -export-symbols-regex "^pam_sm_"
 
-super-clean: maintainer-clean
+clean-all: maintainer-clean
 	rm -fr aclocal autom4te.cache/ m4 missing libtool config.guess
 	rm -fr config.lt config.status config.sub configure depcomp
 	rm -fr libtool install-sh *~ Makefile aclocal.m4 config.h.in ltmain.sh
