+++ src/rest.c
@@ -10,9 +10,48 @@
 #define HOST "login.microsoftonline.com"
 #define PORT "443"
 
+/*
+ * Function: fill_json_buffer
+ * --------------------------
+ * *json_buf: char array that will hold the json message contained in raw_response
+ * 
+ * *raw_response: char array that holds the original raw http response from microsoft. 
+ * 
+ * *start holds the index of the '{' 
+ *
+ * *end holds the index of '}'
+*/
+
 int fill_json_buffer(char *json_buf, char *raw_response, int *start,
-		     int *end);
-int find_json_bounds(char *json_buf, int *start, int *end);
+		     int *end)
+{
+    memcpy(json_buf, &raw_response[*start], *end - *start + 1);
+    json_buf[*end + 1] = '\0';
+    return EXIT_SUCCESS;
+}
+
+/*
+ *
+ * Function: find_json_bounds
+ *---------------------------
+ * json_buf: contains a full http response that includes json.
+ * 
+ * start:    pointer that will contain the index of the start of the json body. 
+ *
+ * end:      pointer that will contain the index of the end of the json body.
+ */
+
+int find_json_bounds(char *json_buf, int *start, int *end)
+{
+    int i, j;
+    for (i = 0; json_buf[i] != '{'; i++) {
+    }
+    *start = i;
+    for (j = i; json_buf[j] != '\0'; j++) {
+    }
+    *end = j;
+    return EXIT_SUCCESS;
+}
 
 /*
  * Function: poll_microsoft_for_token
@@ -22,8 +61,6 @@ int find_json_bounds(char *json_buf, int *start, int *end);
  * *client_id: char array containing the client id 
  *
  * *response_buf: empty buffer to include the response in.
- *
- * returns 0 if completion is successful, 1 if it fails.
  */
 
 int poll_microsoft_for_token(char *code, const char *client_id,
@@ -66,7 +103,7 @@ int poll_microsoft_for_token(char *code, const char *client_id,
 
     if (BIO_do_connect(bio) <= 0) {
 	printf("Failed connection\n");
-	return 1;
+	return EXIT_FAILURE;
     } else {
 	printf("Connected\n");
     }
@@ -114,84 +151,7 @@ int poll_microsoft_for_token(char *code, const char *client_id,
     BIO_free_all(bio);
     SSL_CTX_free(ctx);
 
-    return 0;
-}
-
-/*
- * Function: request_azure_oauth_token
- * -----------------------------------
- * *code: char array containing the device code used in the user's prompt.
- *
- * *client_id: char array containing the client id 
- *
- * returns a 0 if the function completes successfully, and 0 otherwise.
- */
-int request_azure_oauth_token(char *code, const char *client_id,
-			      char *token_buf)
-{
-    int start, end;
-    char response_buf[9000];
-    char json_buf[9000];
-    cJSON *json, *access_token;
-    poll_microsoft_for_token(code, client_id, response_buf);
-    find_json_bounds(response_buf, &start, &end);
-    fill_json_buffer(json_buf, response_buf, &start, &end);
-    json = cJSON_Parse(json_buf);
-    access_token = cJSON_GetObjectItem(json, "access_token");
-    if (access_token == NULL) {
-	/* Something failed. */
-	strcpy(token_buf, "FAILURE");
-	return 1;
-    }
-    strcpy(token_buf,
-	   cJSON_GetObjectItem(json, "access_token")->valuestring);
-    return 0;
-}
-
-/*
- * Function: fill_json_buffer
- * --------------------------
- * *json_buf: char array that will hold the json message contained in raw_response
- * 
- * *raw_response: char array that holds the original raw http response from microsoft. 
- * 
- * *start holds the index of the '{' 
- *
- * *end holds the index of '}'
- *
-*/
-
-int fill_json_buffer(char *json_buf, char *raw_response, int *start,
-		     int *end)
-{
-    memcpy(json_buf, &raw_response[*start], *end - *start + 1);
-    json_buf[*end + 1] = '\0';
-    return 0;
-}
-
-/*
- *
- * Function: find_json_bounds
- *---------------------------
- * json_buf: contains a full http response that includes json.
- * 
- * start:    pointer that will contain the index of the start of the json body. 
- *
- * end:      pointer that will contain the index of the end of the json body.
- *
- * returns an integer that indicates that the function successfully completed. 
- */
-
-int find_json_bounds(char *json_buf, int *start, int *end)
-{
-    int i, j;
-    for (i = 0; json_buf[i] != '{'; i++) {
-    }
-    *start = i;
-    for (j = i; json_buf[j] != '\0'; j++) {
-    }
-    *end = j;
-    return 0;
+    return EXIT_SUCCESS;
 }
 
 int read_code_from_microsoft(const char *client_id,
@@ -234,7 +194,7 @@ int read_code_from_microsoft(const char *client_id,
 
     if (BIO_do_connect(bio) <= 0) {
 	printf("Failed connection\n");
-	return 1;
+	return EXIT_FAILURE;
     } else {
 	printf("Connected\n");
     }
@@ -280,13 +240,13 @@ int read_code_from_microsoft(const char *client_id,
     BIO_free_all(bio);
     SSL_CTX_free(ctx);
 
-    return 0;
+    return EXIT_SUCCESS;
 }
 
 
 int get_microsoft_graph_groups(char *user_object_id, char *response_buf,
 			       char *token, const char *tenant,
-			       char *group_object_id)
+			       const char *group_object_id)
 {
     /* initialize variables */
     char secondary_buf[204800];
@@ -323,7 +283,7 @@ int get_microsoft_graph_groups(char *user_object_id, char *response_buf,
 
     if (BIO_do_connect(bio) <= 0) {
 	printf("Failed connection\n");
-	return 1;
+	return EXIT_FAILURE;
     } else {
 	printf("Connected\n");
     }
@@ -352,7 +312,7 @@ int get_microsoft_graph_groups(char *user_object_id, char *response_buf,
     strcat(secondary_buf, "\"\n");
     strcat(secondary_buf, "\r\n}\r\n\r\n\r\n");
 
-/* Attempts to write len bytes from buf to BIO */
+    /* Attempts to write len bytes from buf to BIO */
     if (BIO_write(bio, secondary_buf, strlen(secondary_buf)) <= 0) {
 	/* handle failed write here */
 	if (!BIO_should_retry(bio)) {
@@ -375,7 +335,7 @@ int get_microsoft_graph_groups(char *user_object_id, char *response_buf,
     }
     BIO_free_all(bio);
     SSL_CTX_free(ctx);
-    return 0;
+    return EXIT_SUCCESS;
 }
 
 int get_microsoft_graph_userprofile(char *token, char *response_buf,
@@ -415,7 +375,7 @@ int get_microsoft_graph_userprofile(char *token, char *response_buf,
 
     if (BIO_do_connect(bio) <= 0) {
 	printf("Failed connection\n");
-	return 1;
+	return EXIT_FAILURE;
     } else {
 	printf("Connected\n");
     }
@@ -456,103 +416,5 @@ int get_microsoft_graph_userprofile(char *token, char *response_buf,
     strcpy(response_buf, "\0");
     BIO_free_all(bio);
     SSL_CTX_free(ctx);
-    return 0;
-}
-
-int parse_user_groups(char *response_buf, cJSON * group_membership_value)
-{
-    char json_buf[4000];
-    int start, end;
-    cJSON *json;
-    strcat(response_buf, "\0");
-    find_json_bounds(response_buf, &start, &end);
-    fill_json_buffer(json_buf, response_buf, &start, &end);
-    json = cJSON_Parse(json_buf);
-    if (json == NULL) {
-	return 1;
-    }
-    int checkval = cJSON_GetObjectItem(json, "value")->type;
-    return checkval;
-}
-
-
-int parse_user_object_id(char *response_buf, char *user_object_id_buf)
-{
-    char json_buf[4000];
-    int start, end;
-    cJSON *json;
-
-    find_json_bounds(response_buf, &start, &end);
-
-    fill_json_buffer(json_buf, response_buf, &start, &end);
-    json = cJSON_Parse(json_buf);
-    cJSON *object = cJSON_GetObjectItem(json, "objectId");
-    if (object == NULL) {
-	/* Something failed. */
-	printf("Something failed.\n");
-	return 1;
-    }
-    strcpy(user_object_id_buf,
-	   cJSON_GetObjectItem(json, "objectId")->valuestring);
-
-    return 0;
-}
-
-/*
- * Function: request_azure_signin_code
- *-----------------------------------
- * *code: character buffer that will have the code inside of it by the function's end.
- *
- * *clientid: contains client id of application as registered with Azure.
- *
- * *tenant: the MS tenant. 
- *
- * returns EXIT_FAILURE if the code buffer is empty at the end of the function
- * and EXIT_SUCCESS if the code buffer is anything but empty. 
- * 
- * TODO: Improve checking if this function succeeded. Should be some more error 
- * handling and there will need to be some way to log failures. 
-*/
-
-int request_azure_signin_code(char *user_code,
-			      const char *client_id, const char *tenant,
-			      char *device_code)
-{
-    char response_buf[2048];
-    char json_buf[2048];
-    cJSON *json;
-    int start, end;
-
-    read_code_from_microsoft(client_id, tenant, response_buf);
-    find_json_bounds(response_buf, &start, &end);
-    fill_json_buffer(json_buf, response_buf, &start, &end);
-    json = cJSON_Parse(json_buf);
-    strcpy(user_code, cJSON_GetObjectItem(json, "user_code")->valuestring);
-    strcpy(device_code,
-	   cJSON_GetObjectItem(json, "device_code")->valuestring);
-    if (user_code[0] == '\0' || device_code[0] == '\0') {
-	/* string is empty, we have failed somewhere */
-	return EXIT_FAILURE;
-    }
     return EXIT_SUCCESS;
-}
-
-int request_azure_group_membership(char *token, char *required_group_id,
-				   const char *tenant)
-{
-//    int success = 2;
-    char user_profile_buf[1000];
-    char user_object_id_buf[100];
-    char raw_group_buf[10000];
-//    cJSON *group_membership_value; 
-
-    get_microsoft_graph_userprofile(token, user_profile_buf, tenant);
-    parse_user_object_id(user_profile_buf, user_object_id_buf);
-    get_microsoft_graph_groups(user_object_id_buf, raw_group_buf, token,
-			       tenant, required_group_id);
-//     int is_in_group = parse_user_groups(raw_group_buf, group_membership_value);
-//     if(is_in_group == success){
-//         return 0;
-//     }
-    return 1;
 }
